Move actions, button, label, modal and spinner components in common
