+++ libs/search-widget/src/common/button/Button.scss
