+++ libs/search-widget/src/common/button/Button.svelte
