+++ libs/search-widget/src/common/button/button.spec.ts
