+++ libs/search-widget/src/common/label/Label.svelte
