+++ libs/search-widget/src/common/modal/Modal.scss
