+++ libs/search-widget/src/common/modal/Modal.svelte
