+++ libs/search-widget/src/common/modal/ModalHeader.scss
