+++ libs/search-widget/src/common/modal/modal.spec.ts
