+++ libs/search-widget/src/common/spinner/LoadingDots.scss
