+++ libs/search-widget/src/common/spinner/LoadingDots.svelte
