+++ libs/search-widget/src/common/spinner/Spinner.scss
