+++ libs/search-widget/src/common/spinner/Spinner.svelte
