+++ libs/search-widget/src/viewer/menus/ActionMenu.svelte
@@ -2,7 +2,7 @@
   import type { WidgetAction } from '../../core/models';
   import { getWidgetActions } from '../../core/store';
   import { getCDN } from '../../core/utils';
-  import { clickOutside } from '../../components/actions/actions';
+  import { clickOutside } from '../../common/actions/actions';
 
   export let uid: string;
